requirements.txt
playwright>=1.40.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.90.0
pandas>=2.0.0
streamlit>=1.32.0
plotly>=5.20.0
